// src/app/generated/alloy.api/model/eventTemplateRole.ts
/**
 * Alloy API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EventTemplatePermission } from './eventTemplatePermission';


export interface EventTemplateRole { 
    id?: string;
    name?: string | null;
    allPermissions?: boolean;
    permissions?: Array<EventTemplatePermission> | null;
}